+++ kernel/src/FatFS.h
@@ -32,7 +32,7 @@ public:
         else {
             // Read from device failed.
             srl.writestr("[FatFS]: ERROR -> Could not read from device at port ");
-            srl.writestr(to_string((u64)portNumber));
+            srl.writestr(to_string(portNumber));
             srl.writestr("\r\n");
         }
     }
